moved serialization tests to layer classes
