+++ keras/src/quantizers/awq_test.py
@@ -707,131 +707,3 @@ class AWQAccuracyTest(testing.TestCase):
         )
 
         awq_obj.free()
-
-
-def _create_dense_layer():
-    """Helper to create a Dense layer for testing."""
-    input_shape = (None, 8)
-    layer = layers.Dense(units=16)
-    layer.build(input_shape)
-    return layer, layers.Dense, input_shape
-
-
-def _create_einsum_dense_layer():
-    """Helper to create an EinsumDense layer for testing."""
-    input_shape = (None, 3)
-    layer = layers.EinsumDense(
-        equation="ab,bcd->acd", output_shape=(8, 32), bias_axes="d"
-    )
-    layer.build(input_shape)
-    return layer, layers.EinsumDense, input_shape
-
-
-@pytest.mark.requires_trainable_backend
-class AWQSerializationTest(testing.TestCase):
-    """Tests for AWQ layer serialization and deserialization."""
-
-    @parameterized.named_parameters(
-        ("dense", _create_dense_layer),
-        ("einsum_dense", _create_einsum_dense_layer),
-    )
-    def test_awq_layer_serialization(self, layer_factory):
-        """Test that an AWQ-quantized layer can be serialized and
-        deserialized correctly."""
-        layer, layer_class, input_shape = layer_factory()
-
-        layer.quantize(
-            "awq",
-            config=AWQConfig(
-                dataset=None, tokenizer=None, group_size=8, n_grid=10
-            ),
-        )
-        config = layer.get_config()
-        new_layer = layer_class.from_config(config)
-        new_layer.build(input_shape)
-        self.assertEqual(new_layer.quantization_mode, "awq")
-
-    def _assert_awq_variables_loaded(self, layer, awq_store):
-        """Helper to verify AWQ variables are loaded correctly."""
-        self.assertTrue(layer.is_awq_calibrated)
-        self.assertAllClose(layer.bias, awq_store["0"])
-        self.assertAllClose(layer.quantized_kernel, awq_store["1"])
-        self.assertAllClose(layer.kernel_scale, awq_store["2"])
-        self.assertAllClose(layer.kernel_zero, awq_store["3"])
-        self.assertAllClose(layer.awq_scales, awq_store["4"])
-        self.assertAllClose(layer.g_idx, awq_store["5"])
-
-    def test_awq_dense_legacy_load_own_variables(self):
-        """Test loading AWQ variables from a legacy store format."""
-        awq_store = {
-            "0": np.random.random((16,)).astype("float32"),  # bias
-            "1": np.random.randint(0, 16, size=(8, 8), dtype="uint8"),  # kernel
-            "2": np.random.random((16, 1)).astype("float32"),  # scale
-            "3": np.random.random((16, 1)).astype("uint8"),  # zero
-            "4": np.random.random((8,)).astype("float32"),  # awq_scales
-            "5": np.random.random((8,)).astype("float32"),  # g_idx
-        }
-        layer = layers.Dense(units=16, dtype="awq/4/8_from_float32")
-        layer.build((None, 8))
-        layer.load_own_variables(awq_store)
-        self._assert_awq_variables_loaded(layer, awq_store)
-
-    def test_awq_einsum_dense_legacy_load_own_variables(self):
-        """Test loading AWQ variables from a legacy store format for
-        EinsumDense."""
-        # kernel shape (3, 8, 32), packed: (16, 24) for 4-bit
-        awq_store = {
-            "0": np.random.random((32,)).astype("float32"),  # bias
-            "1": np.random.randint(0, 16, size=(16, 24), dtype="uint8"),
-            "2": np.random.random((32, 3)).astype("float32"),  # scale
-            "3": np.random.random((32, 3)).astype("uint8"),  # zero
-            "4": np.random.random((24,)).astype("float32"),  # awq_scales
-            "5": np.random.random((24,)).astype("float32"),  # g_idx
-        }
-        layer = layers.EinsumDense(
-            equation="ab,bcd->acd",
-            output_shape=(8, 32),
-            bias_axes="d",
-            dtype="awq/4/8_from_float32",
-        )
-        layer.build((None, 3))
-        layer.load_own_variables(awq_store)
-        self._assert_awq_variables_loaded(layer, awq_store)
-
-    def test_int4_awq_kernel_returns_unpacked_form(self):
-        """Test that the `kernel` property returns the unpacked int4 AWQ
-        kernel."""
-        from keras.src import quantizers
-
-        layer = layers.Dense(units=2)
-        layer.build((None, 2))
-        layer.quantize(
-            "awq",
-            config=AWQConfig(
-                dataset=None, tokenizer=None, group_size=8, n_grid=10
-            ),
-        )
-        layer.is_awq_calibrated = True  # Bypass calibration check
-        packed_kernel = layer.quantized_kernel
-        self.assertAllClose(
-            layer.kernel, quantizers.unpack_int4(packed_kernel, 2)
-        )
-
-    def test_awq_kernel_packing(self):
-        """Validates that 4-bit AWQ packing reduces the kernel size."""
-        layer = layers.Dense(units=16, use_bias=False)
-        layer.build((None, 8))
-
-        original_kernel_params = ops.prod(layer._kernel.shape)
-
-        layer.quantize(
-            "awq",
-            config=AWQConfig(
-                dataset=None, tokenizer=None, group_size=8, n_grid=10
-            ),
-        )
-
-        quantized_kernel_params = ops.prod(layer.quantized_kernel.shape)
-        self.assertEqual(
-            quantized_kernel_params, ops.floor_divide(original_kernel_params, 2)
-        )
